+++ init.lua
@@ -317,12 +317,12 @@ require('lazy').setup({
   },
 
   require 'kickstart.plugins.autocomplete',
-  require 'kickstart.plugins.tokyonight',
+  require 'kickstart.plugins.themes.tokyonight',
   -- Highlight todo, notes, etc in comments
   { 'folke/todo-comments.nvim', event = 'VimEnter', dependencies = { 'nvim-lua/plenary.nvim' }, opts = { signs = false } },
   require 'kickstart.plugins.mini',
   require 'kickstart.plugins.treesitter',
-  
+
   -- The following two comments only work if you have downloaded the kickstart repo, not just copy pasted the
   -- init.lua. If you want these files, they are in the repository, so you can just download them and
   -- place them in the correct locations.
